test(vket): 巨大テストを責務別に分割
Vketテストの振る舞いを保ったまま、責務単位で読めるサイズに整理するため。

検証: Vket全148件、ruff 0.11.6、file-size-guard
