--- app/vket/tests/test_vket.py
@@ -1,3401 +1,0 @@
-"""Vketコラボ機能のテスト."""
-
-from datetime import time, timedelta
-from importlib import import_module
-
-from django.contrib.auth import get_user_model
-from django.apps import apps
-from django.core.cache import cache
-from django.core.exceptions import ValidationError
-from django.test import Client, TestCase
-from django.urls import reverse
-from django.utils import timezone
-
-from community.models import Community, CommunityMember
-from event.models import Event, EventDetail, EventOccurrenceTombstone
-from ta_hub.index_cache import get_index_view_cache_key
-from vket.models import (
-    VketCollaboration,
-    VketNotice,
-    VketNoticeReceipt,
-    VketParticipation,
-    VketPresentation,
-)
-from vket.services import sync_participation_publication
-from vket.views.helpers import _build_schedule_context
-
-
-User = get_user_model()
-
-
-class VketPublicPagesTests(TestCase):
-    def setUp(self):
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-summer',
-            name='Vket 2026 Summer 技術学術WEEK',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-            hashtags=['#Vketステージ', '#Vket技術学術WEEK'],
-        )
-        self.draft_collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-draft',
-            name='Vket Draft',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.DRAFT,
-        )
-
-    def test_list_page(self):
-        client = Client()
-        response = client.get(reverse('vket:list'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.collaboration.name)
-        # 下書きは表示されない
-        self.assertNotContains(response, self.draft_collaboration.name)
-
-    def test_detail_page(self):
-        client = Client()
-        response = client.get(reverse('vket:detail', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.collaboration.name)
-
-    def test_detail_page_is_404_for_draft(self):
-        client = Client()
-        response = client.get(reverse('vket:detail', kwargs={'pk': self.draft_collaboration.pk}))
-        self.assertEqual(response.status_code, 404)
-
-    def test_collaboration_validation(self):
-        today = timezone.localdate()
-        collaboration = VketCollaboration(
-            slug='invalid-collab',
-            name='Invalid collaboration',
-            period_start=today + timedelta(days=1),
-            period_end=today,
-            registration_deadline=today,
-            lt_deadline=today - timedelta(days=1),
-            phase=VketCollaboration.Phase.DRAFT,
-        )
-        with self.assertRaises(ValidationError):
-            collaboration.full_clean()
-
-
-class VketApplyFlowTests(TestCase):
-    def setUp(self):
-        self.client = Client()
-        self.owner = User.objects.create_user(
-            user_name='owner_user',
-            email='owner@example.com',
-            password='testpass123',
-        )
-        self.other_user = User.objects.create_user(
-            user_name='other_user',
-            email='other@example.com',
-            password='testpass123',
-        )
-
-        self.community = Community.objects.create(
-            name='個人開発集会',
-            status='approved',
-            frequency='毎週',
-        )
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.owner,
-            role=CommunityMember.Role.OWNER,
-        )
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.other_user,
-            role=CommunityMember.Role.STAFF,
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-apply-test',
-            name='Vket 2026 Summer 技術学術WEEK',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-        )
-
-        # 日付選択肢の元になるイベントを作成
-        Event.objects.create(
-            community=self.community,
-            date=today,
-            start_time='22:00',
-            duration=60,
-        )
-
-    def _set_active_community(self):
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-
-    def test_apply_get_allows_staff(self):
-        """スタッフもapplyページにアクセスできる"""
-        self.client.login(username='other_user', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-
-    def test_apply_get_shows_organizer_note_guidance_above_textarea(self):
-        """備考欄の案内文がテキストエリア外に表示される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(
-            response,
-            'スタッフが少なく、当日ステージへの案内などサポートが欲しい場合は「サポート希望」と記載してください。',
-        )
-        self.assertContains(
-            response,
-            'その他、運営に連絡事項があれば記載をお願いします。',
-        )
-        self.assertContains(response, 'background: #eef6ff; color: #2f5f8f;')
-        self.assertContains(response, 'text-decoration-underline')
-
-    def test_apply_post_creates_participation_and_presentation(self):
-        """主催者が申請するとVketParticipationとVketPresentationが作成される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        target_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data={
-                'requested_date': target_date.isoformat(),
-                'requested_start_time': '21:00',
-                'requested_duration': '60',
-                'organizer_note': '備考テスト',
-                # formset management form
-                'lt-TOTAL_FORMS': '1',
-                'lt-INITIAL_FORMS': '0',
-                'lt-MIN_NUM_FORMS': '0',
-                'lt-MAX_NUM_FORMS': '20',
-                # LT data
-                'lt-0-speaker': 'テスト登壇者',
-                'lt-0-theme': 'テストテーマ',
-                'lt-0-lt_start_time': '',
-            },
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-
-        participation = VketParticipation.objects.get(
-            collaboration=self.collaboration, community=self.community
-        )
-        # 希望日程が保存されている
-        self.assertEqual(participation.requested_date, target_date)
-        self.assertEqual(participation.requested_start_time.strftime('%H:%M'), '21:00')
-        self.assertEqual(participation.requested_duration, 60)
-        self.assertEqual(participation.organizer_note, '備考テスト')
-        # applied_byがセットされている
-        self.assertEqual(participation.applied_by, self.owner)
-        self.assertIsNotNone(participation.applied_at)
-        self.assertEqual(participation.progress, VketParticipation.Progress.APPLIED)
-
-        # 確定前なのでEventは作成されない
-        self.assertIsNone(participation.published_event_id)
-
-        # VketPresentationが作成されている
-        pres = VketPresentation.objects.get(participation=participation, order=0)
-        self.assertEqual(pres.speaker, 'テスト登壇者')
-        self.assertEqual(pres.theme, 'テストテーマ')
-
-    def test_apply_is_forbidden_after_registration_deadline_for_new_participation(self):
-        """参加申請締切後は新規参加登録が403になる"""
-        self.collaboration.registration_deadline = timezone.localdate() - timedelta(days=1)
-        self.collaboration.save()
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 403)
-
-    def test_detail_can_apply_is_false_when_registration_closed_and_no_participation(self):
-        """締切後・未参加の場合 can_apply=False"""
-        self.collaboration.registration_deadline = timezone.localdate() - timedelta(days=1)
-        self.collaboration.save()
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(reverse('vket:detail', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertFalse(response.context['can_apply'])
-
-    def test_detail_can_apply_is_true_when_lt_open_and_participation_has_published_event(self):
-        """LT締切内かつpublished_eventがある場合 can_apply=True"""
-        today = timezone.localdate()
-        self.collaboration.registration_deadline = today - timedelta(days=1)
-        self.collaboration.lt_deadline = today + timedelta(days=1)
-        self.collaboration.phase = VketCollaboration.Phase.SCHEDULING
-        self.collaboration.save()
-
-        weekday_code = ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'][today.weekday()]
-        event = Event.objects.create(
-            community=self.community,
-            date=today,
-            start_time='21:00',
-            duration=60,
-            weekday=weekday_code,
-        )
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            published_event=event,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(reverse('vket:detail', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.context['can_apply'])
-
-    def test_apply_accepts_late_lt_submission_as_draft_for_existing_participation(self):
-        """発表締切後も既存参加団体の発表情報は申請中で保存できる"""
-        today = timezone.localdate()
-        self.collaboration.registration_deadline = today - timedelta(days=3)
-        self.collaboration.lt_deadline = today - timedelta(days=1)
-        self.collaboration.phase = VketCollaboration.Phase.ANNOUNCEMENT
-        self.collaboration.save()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            progress=VketParticipation.Progress.APPLIED,
-            applied_by=self.owner,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.context['permissions'].can_edit_lt)
-        self.assertTrue(response.context['is_late_lt_submission'])
-        self.assertContains(response, '運営の確認後に確定・公開されます。')
-
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '21:00',
-            'requested_duration': '60',
-            'organizer_note': '締切後の追記',
-        }
-        post_data.update(
-            self._make_formset_data([
-                {'speaker': '締切後登壇者', 'theme': '締切後テーマ'}
-            ])
-        )
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        participation.refresh_from_db()
-        presentation = VketPresentation.objects.get(participation=participation)
-        self.assertEqual(participation.organizer_note, '締切後の追記')
-        self.assertEqual(presentation.speaker, '締切後登壇者')
-        self.assertEqual(presentation.status, VketPresentation.Status.DRAFT)
-
-    def test_late_lt_update_resets_confirmed_presentation_to_draft(self):
-        """締切後に確定済み発表を更新した場合は申請中に戻す"""
-        today = timezone.localdate()
-        self.collaboration.registration_deadline = today - timedelta(days=3)
-        self.collaboration.lt_deadline = today - timedelta(days=1)
-        self.collaboration.phase = VketCollaboration.Phase.ANNOUNCEMENT
-        self.collaboration.save()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-            progress=VketParticipation.Progress.REHEARSAL,
-            applied_by=self.owner,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='確定済み登壇者',
-            theme='確定済みテーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '23:00',
-            'requested_duration': '90',
-            'organizer_note': '締切後更新',
-        }
-        post_data.update(
-            self._make_formset_data(
-                [
-                    {
-                        'speaker': '更新後登壇者',
-                        'theme': '更新後テーマ',
-                        'lt_start_time': '23:30',
-                    },
-                ],
-                initial_forms=1,
-            )
-        )
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        presentation.refresh_from_db()
-        self.assertEqual(presentation.speaker, '更新後登壇者')
-        self.assertEqual(presentation.theme, '更新後テーマ')
-        self.assertEqual(presentation.requested_start_time.strftime('%H:%M'), '21:30')
-        self.assertEqual(presentation.status, VketPresentation.Status.DRAFT)
-
-    def test_apply_blocks_lt_submission_after_event_period(self):
-        """開催期間後は発表情報も編集不可にする"""
-        today = timezone.localdate()
-        self.collaboration.period_start = today - timedelta(days=8)
-        self.collaboration.period_end = today - timedelta(days=1)
-        self.collaboration.registration_deadline = today - timedelta(days=7)
-        self.collaboration.lt_deadline = today - timedelta(days=2)
-        self.collaboration.phase = VketCollaboration.Phase.ANNOUNCEMENT
-        self.collaboration.save()
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            progress=VketParticipation.Progress.APPLIED,
-            applied_by=self.owner,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertFalse(response.context['permissions'].can_edit_lt)
-        self.assertFalse(response.context['is_late_lt_submission'])
-        self.assertContains(response, '発表情報（Step 2）は受付期間外のため編集できません。')
-
-    def test_status_page_shows_register_complete_guidance(self):
-        """参加状況画面で登録後の次アクションが明示される"""
-        self.collaboration.slug = 'vket-2026-summer'
-        self.collaboration.save(update_fields=['slug'])
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            applied_by=self.owner,
-            applied_at=timezone.now(),
-            progress=VketParticipation.Progress.APPLIED,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(reverse('vket:status', kwargs={'pk': self.collaboration.pk}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(
-            response,
-            'Vket公式サイトのステージページに、イベント情報を暫定内容で登録してください。',
-        )
-        self.assertContains(
-            response,
-            '日程や発表内容が未確定でも、後から更新できます。',
-        )
-        self.assertContains(
-            response,
-            '登録が終わったら、この画面に戻って「登録完了」を押してください。',
-        )
-        self.assertContains(
-            response,
-            '開催会場は「Parareal Central Ignition Point - 着火点 - - エントランス」に設定してください。',
-        )
-        self.assertContains(response, 'タグは「Vketステージ」を選択してください。')
-        self.assertContains(
-            response,
-            'Vket側で登録しただけでは Hub の進捗は更新されません。',
-        )
-        self.assertContains(
-            response,
-            'https://vket.com/hub/2026Summer/notification',
-        )
-        self.assertContains(response, 'Vketステージに登録する')
-        self.assertNotContains(response, '参加申込みが完了しました。')
-
-        body = response.content.decode()
-        self.assertLess(
-            body.index('Vketステージに登録する'),
-            body.index('<i class="fas fa-check me-1"></i>登録完了'),
-        )
-
-    def test_status_page_uses_collaboration_stage_url_when_configured(self):
-        """コラボに登録URLが設定されている場合はそのURLを使う"""
-        custom_stage_url = 'https://example.com/vket/custom-stage'
-        self.collaboration.settings_json = {'stage_url': f' {custom_stage_url} '}
-        self.collaboration.save(update_fields=['settings_json'])
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            applied_by=self.owner,
-            applied_at=timezone.now(),
-            progress=VketParticipation.Progress.APPLIED,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(reverse('vket:status', kwargs={'pk': self.collaboration.pk}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, custom_stage_url)
-        self.assertNotContains(response, 'https://vket.com/hub/2026Summer/notification')
-        self.assertNotContains(response, 'Parareal Central Ignition Point - 着火点 - - エントランス')
-
-    def test_lt_start_time_saved_to_presentation(self):
-        """LT開始時刻が VketPresentation.requested_start_time に保存される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        target_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data={
-                'requested_date': target_date.isoformat(),
-                'requested_start_time': '21:00',
-                'requested_duration': '60',
-                'organizer_note': '備考テスト',
-                'lt-TOTAL_FORMS': '1',
-                'lt-INITIAL_FORMS': '0',
-                'lt-MIN_NUM_FORMS': '0',
-                'lt-MAX_NUM_FORMS': '20',
-                'lt-0-speaker': 'テスト登壇者',
-                'lt-0-theme': 'テストテーマ',
-                'lt-0-lt_start_time': '21:30',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        participation = VketParticipation.objects.get(
-            collaboration=self.collaboration, community=self.community
-        )
-        pres = VketPresentation.objects.get(participation=participation, order=0)
-        self.assertEqual(pres.requested_start_time.strftime('%H:%M'), '21:30')
-
-    def test_schedule_table_uses_unpublished_presentation_start_time(self):
-        """未公開の発表開始時刻が日程表のLTマーカーに反映される"""
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            progress=VketParticipation.Progress.STAGE_REGISTERED,
-        )
-        VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='未定',
-            theme='未定',
-            requested_start_time=time(21, 30),
-            duration=30,
-        )
-
-        context = _build_schedule_context(self.collaboration, include_requested=True)
-        row = next(
-            r for r in context['rows'] if r['participation'].pk == participation.pk
-        )
-        lt_tooltips = [
-            cell['lt_tooltip'] for cell in row['cells'] if cell['lt_times']
-        ]
-
-        self.assertEqual(row['start_time'], time(21, 0))
-        self.assertEqual(lt_tooltips, ['21:30'])
-
-    def test_new_apply_shows_organizer_note_template(self):
-        """新規申請GETで organizer_note の初期値テンプレートが表示される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        response = self.client.get(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        form = response.context['form']
-        self.assertIn('当日サポートが欲しい', form.initial.get('organizer_note', ''))
-
-    def test_existing_participation_preserves_organizer_note(self):
-        """既存参加者のGETで organizer_note が初期テンプレートで上書きされない"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        # 先に参加を作成（副作用でDBレコードを作成）
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            organizer_note='カスタム備考',
-            progress=VketParticipation.Progress.APPLIED,
-        )
-
-        response = self.client.get(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        form = response.context['form']
-        self.assertEqual(form.initial.get('organizer_note'), 'カスタム備考')
-
-    def _make_formset_data(self, lt_rows, initial_forms=0):
-        """formset用のPOSTデータを辞書で返すヘルパー"""
-        data = {
-            'lt-TOTAL_FORMS': str(len(lt_rows)),
-            'lt-INITIAL_FORMS': str(initial_forms),
-            'lt-MIN_NUM_FORMS': '0',
-            'lt-MAX_NUM_FORMS': '20',
-        }
-        for i, row in enumerate(lt_rows):
-            data[f'lt-{i}-speaker'] = row.get('speaker', '')
-            data[f'lt-{i}-theme'] = row.get('theme', '')
-            data[f'lt-{i}-lt_start_time'] = row.get('lt_start_time', '')
-            if row.get('DELETE'):
-                data[f'lt-{i}-DELETE'] = 'on'
-        return data
-
-    def test_apply_creates_multiple_presentations(self):
-        """複数LTを送信するとDBに複数件作成される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        target_date = self.collaboration.period_start
-        lt_rows = [
-            {'speaker': '登壇者A', 'theme': 'テーマA'},
-            {'speaker': '登壇者B', 'theme': 'テーマB'},
-            {'speaker': '登壇者C', 'theme': 'テーマC'},
-        ]
-        post_data = {
-            'requested_date': target_date.isoformat(),
-            'requested_start_time': '21:00',
-            'requested_duration': '60',
-            'organizer_note': '',
-        }
-        post_data.update(self._make_formset_data(lt_rows))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        participation = VketParticipation.objects.get(
-            collaboration=self.collaboration, community=self.community
-        )
-        presentations = list(
-            VketPresentation.objects.filter(participation=participation).order_by('order')
-        )
-        self.assertEqual(len(presentations), 3)
-        self.assertEqual(presentations[0].speaker, '登壇者A')
-        self.assertEqual(presentations[0].order, 0)
-        self.assertEqual(presentations[1].speaker, '登壇者B')
-        self.assertEqual(presentations[1].order, 1)
-        self.assertEqual(presentations[2].speaker, '登壇者C')
-        self.assertEqual(presentations[2].order, 2)
-
-    def test_apply_deletes_presentation(self):
-        """DELETEフラグ付きで送信すると該当レコードが削除される"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        # 先に参加と2件のプレゼンを作成
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time='21:00',
-            requested_duration=60,
-            progress=VketParticipation.Progress.APPLIED,
-            applied_by=self.owner,
-        )
-        VketPresentation.objects.create(
-            participation=participation, order=0, speaker='残す登壇者', theme='残すテーマ'
-        )
-        VketPresentation.objects.create(
-            participation=participation, order=1, speaker='消す登壇者', theme='消すテーマ'
-        )
-
-        target_date = self.collaboration.period_start
-        lt_rows = [
-            {'speaker': '残す登壇者', 'theme': '残すテーマ'},
-            {'speaker': '消す登壇者', 'theme': '消すテーマ', 'DELETE': True},
-        ]
-        post_data = {
-            'requested_date': target_date.isoformat(),
-            'requested_start_time': '21:00',
-            'requested_duration': '60',
-            'organizer_note': '',
-        }
-        post_data.update(self._make_formset_data(lt_rows, initial_forms=2))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        presentations = list(
-            VketPresentation.objects.filter(participation=participation).order_by('order')
-        )
-        self.assertEqual(len(presentations), 1)
-        self.assertEqual(presentations[0].speaker, '残す登壇者')
-
-    def test_confirmed_participation_post_allows_unlocked_lt_start_time(self):
-        """日程確定後も締切内なら未確定LTの開始時刻を更新できる"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-            progress=VketParticipation.Progress.REHEARSAL,
-            applied_by=self.owner,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='確定前登壇者',
-            theme='確定前テーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.DRAFT,
-        )
-
-        post_data = {
-            'requested_date': (self.collaboration.period_start + timedelta(days=1)).isoformat(),
-            'requested_start_time': '23:00',
-            'requested_duration': '90',
-            'organizer_note': '確定後も備考は更新',
-            'lt_slot_minutes': '20',
-        }
-        post_data.update(
-            self._make_formset_data(
-                [
-                    {
-                        'speaker': '更新後登壇者',
-                        'theme': '更新後テーマ',
-                        'lt_start_time': '23:30',
-                    },
-                    {
-                        'speaker': '追加登壇者',
-                        'theme': '追加テーマ',
-                        'lt_start_time': '23:50',
-                    },
-                ],
-                initial_forms=1,
-            )
-        )
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        participation.refresh_from_db()
-        presentation.refresh_from_db()
-        self.assertEqual(participation.requested_date, self.collaboration.period_start)
-        self.assertEqual(participation.requested_start_time, time(21, 0))
-        self.assertEqual(participation.requested_duration, 60)
-        self.assertEqual(participation.organizer_note, '確定後も備考は更新')
-        self.assertEqual(participation.lt_slot_minutes, 20)
-        self.assertEqual(presentation.speaker, '更新後登壇者')
-        self.assertEqual(presentation.theme, '更新後テーマ')
-        self.assertEqual(presentation.requested_start_time, time(23, 30))
-        self.assertEqual(
-            VketPresentation.objects.get(participation=participation, order=1).requested_start_time,
-            time(23, 50),
-        )
-
-    def test_confirmed_participation_without_event_can_update_lt_and_note(self):
-        """Eventがない確定済み参加でも発表と備考を更新できる"""
-        Event.objects.filter(community=self.community).delete()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-            progress=VketParticipation.Progress.REHEARSAL,
-            applied_by=self.owner,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='更新前登壇者',
-            theme='更新前テーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '23:00',
-            'requested_duration': '90',
-            'organizer_note': 'Eventなしでも更新できる備考',
-        }
-        post_data.update(
-            self._make_formset_data(
-                [{'speaker': '更新後登壇者', 'theme': '更新後テーマ'}],
-                initial_forms=1,
-            )
-        )
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        participation.refresh_from_db()
-        presentation.refresh_from_db()
-        self.assertEqual(participation.requested_date, self.collaboration.period_start)
-        self.assertEqual(participation.requested_start_time, time(21, 0))
-        self.assertEqual(participation.requested_duration, 60)
-        self.assertEqual(participation.confirmed_date, self.collaboration.period_start)
-        self.assertEqual(participation.confirmed_start_time, time(21, 0))
-        self.assertEqual(participation.confirmed_duration, 60)
-        self.assertEqual(participation.organizer_note, 'Eventなしでも更新できる備考')
-        self.assertEqual(presentation.speaker, '更新後登壇者')
-        self.assertEqual(presentation.theme, '更新後テーマ')
-        self.assertEqual(presentation.requested_start_time, time(21, 30))
-
-    def test_confirmed_participation_post_does_not_delete_existing_lt(self):
-        """日程確定後はformset DELETEでも既存LTを削除しない"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-            progress=VketParticipation.Progress.REHEARSAL,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='削除されない登壇者',
-            theme='削除されないテーマ',
-            requested_start_time=time(21, 30),
-        )
-
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '21:00',
-            'requested_duration': '60',
-            'organizer_note': '',
-        }
-        post_data.update(
-            self._make_formset_data(
-                [
-                    {
-                        'speaker': '削除されない登壇者',
-                        'theme': '削除されないテーマ',
-                        'DELETE': True,
-                    }
-                ],
-                initial_forms=1,
-            )
-        )
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(VketPresentation.objects.filter(pk=presentation.pk).exists())
-
-    def test_confirmed_participation_apply_get_shows_lt_editable_message(self):
-        """日程確定後も締切内のLT時刻編集可否を案内する"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0),
-            requested_duration=60,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-        )
-
-        response = self.client.get(reverse('vket:apply', kwargs={'pk': self.collaboration.pk}))
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, '参加日程（Step 1）は運営が確定済みのため編集できません。')
-        self.assertContains(response, '発表開始時刻は発表情報の締切まで変更できます')
-
-    def test_lt_start_time_is_rejected_after_deadline_but_text_updates(self):
-        """締切後はLT時刻を保持し、登壇者名とテーマは更新できる"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        self.collaboration.lt_deadline = timezone.localdate() - timedelta(days=1)
-        self.collaboration.save(update_fields=['lt_deadline'])
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration, community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0), requested_duration=60,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation, order=0, speaker='変更前登壇者',
-            theme='変更前テーマ', requested_start_time=time(21, 30),
-        )
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '21:00', 'requested_duration': '60',
-            'organizer_note': '', 'lt_slot_minutes': '45',
-        }
-        post_data.update(self._make_formset_data([
-            {'speaker': '変更後登壇者', 'theme': '変更後テーマ', 'lt_start_time': '23:30'},
-        ], initial_forms=1))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}), post_data
-        )
-
-        self.assertEqual(response.status_code, 302)
-        presentation.refresh_from_db()
-        self.assertEqual(presentation.speaker, '変更後登壇者')
-        self.assertEqual(presentation.theme, '変更後テーマ')
-        self.assertEqual(presentation.requested_start_time, time(21, 30))
-
-    def test_confirmed_or_published_lt_time_ignores_post_value(self):
-        """確定済みまたは公開済みLTの時刻はPOST値で上書きできない"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration, community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time=time(21, 0), requested_duration=60,
-        )
-        confirmed = VketPresentation.objects.create(
-            participation=participation, order=0, speaker='確定済み', theme='テーマ',
-            requested_start_time=time(21, 30), status=VketPresentation.Status.CONFIRMED,
-        )
-        published_detail = EventDetail.objects.create(
-            event=Event.objects.get(community=self.community),
-            speaker='公開済み', theme='テーマ', start_time=time(22, 0),
-        )
-        published = VketPresentation.objects.create(
-            participation=participation, order=1, speaker='公開済み', theme='テーマ',
-            requested_start_time=time(22, 0), published_event_detail=published_detail,
-        )
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '21:00', 'requested_duration': '60',
-            'organizer_note': '', 'lt_slot_minutes': '30',
-        }
-        post_data.update(self._make_formset_data([
-            {'speaker': '確定済み更新', 'theme': '更新', 'lt_start_time': '23:00'},
-            {'speaker': '公開済み更新', 'theme': '更新', 'lt_start_time': '23:30'},
-        ], initial_forms=2))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}), post_data
-        )
-
-        self.assertEqual(response.status_code, 302)
-        confirmed.refresh_from_db()
-        published.refresh_from_db()
-        self.assertEqual(confirmed.requested_start_time, time(21, 30))
-        self.assertEqual(published.requested_start_time, time(22, 0))
-
-    def test_missing_lt_start_times_are_assigned_from_slot_minutes(self):
-        """未入力LT時刻は参加枠の開始時刻から持ち時間ごとに補完する"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        post_data = {
-            'requested_date': self.collaboration.period_start.isoformat(),
-            'requested_start_time': '21:00', 'requested_duration': '60',
-            'organizer_note': '', 'lt_slot_minutes': '20',
-        }
-        post_data.update(self._make_formset_data([
-            {'speaker': '登壇者1', 'theme': 'テーマ1'},
-            {'speaker': '登壇者2', 'theme': 'テーマ2'},
-            {'speaker': '登壇者3', 'theme': 'テーマ3'},
-        ]))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}), post_data
-        )
-
-        self.assertEqual(response.status_code, 302)
-        participation = VketParticipation.objects.get(
-            collaboration=self.collaboration, community=self.community
-        )
-        self.assertEqual(participation.lt_slot_minutes, 20)
-        self.assertEqual(
-            list(participation.presentations.values_list('requested_start_time', flat=True)),
-            [time(21, 0), time(21, 20), time(21, 40)],
-        )
-
-    def test_lt_slot_minutes_defaults_to_30(self):
-        """参加枠の持ち時間は30分を既定値にする"""
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration, community=self.community,
-        )
-        self.assertEqual(participation.lt_slot_minutes, 30)
-
-    def test_apply_get_prefills_multiple_presentations(self):
-        """既存の複数LTがGETでformsetにプリフィルされる"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time='21:00',
-            requested_duration=60,
-            progress=VketParticipation.Progress.APPLIED,
-            applied_by=self.owner,
-        )
-        VketPresentation.objects.create(
-            participation=participation, order=0, speaker='登壇者1', theme='テーマ1'
-        )
-        VketPresentation.objects.create(
-            participation=participation, order=1, speaker='登壇者2', theme='テーマ2'
-        )
-        VketPresentation.objects.create(
-            participation=participation, order=2, speaker='登壇者3', theme='テーマ3'
-        )
-
-        response = self.client.get(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        formset = response.context['formset']
-        # initial + extra(1) = 4 forms
-        self.assertEqual(len(formset.forms), 4)
-        self.assertEqual(formset.forms[0].initial['speaker'], '登壇者1')
-        self.assertEqual(formset.forms[1].initial['speaker'], '登壇者2')
-        self.assertEqual(formset.forms[2].initial['speaker'], '登壇者3')
-
-    def test_apply_skips_empty_presentation_rows(self):
-        """空行はスキップされる（DBに保存されない）"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-
-        target_date = self.collaboration.period_start
-        lt_rows = [
-            {'speaker': '登壇者A', 'theme': 'テーマA'},
-            {'speaker': '', 'theme': ''},
-            {'speaker': '登壇者C', 'theme': 'テーマC'},
-        ]
-        post_data = {
-            'requested_date': target_date.isoformat(),
-            'requested_start_time': '21:00',
-            'requested_duration': '60',
-            'organizer_note': '',
-        }
-        post_data.update(self._make_formset_data(lt_rows))
-
-        response = self.client.post(
-            reverse('vket:apply', kwargs={'pk': self.collaboration.pk}),
-            data=post_data,
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        participation = VketParticipation.objects.get(
-            collaboration=self.collaboration, community=self.community
-        )
-        presentations = list(
-            VketPresentation.objects.filter(participation=participation).order_by('order')
-        )
-        self.assertEqual(len(presentations), 2)
-        self.assertEqual(presentations[0].speaker, '登壇者A')
-        self.assertEqual(presentations[0].order, 0)
-        self.assertEqual(presentations[1].speaker, '登壇者C')
-        self.assertEqual(presentations[1].order, 1)
-
-    def test_presentation_delete_by_organizer(self):
-        """主催者がLTを削除できる"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            progress=VketParticipation.Progress.APPLIED,
-        )
-        pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='削除テスト',
-            theme='テーマ',
-            status=VketPresentation.Status.DRAFT,
-        )
-        response = self.client.post(
-            reverse(
-                'vket:presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-        self.assertEqual(response.status_code, 302)
-        self.assertFalse(VketPresentation.objects.filter(pk=pres.pk).exists())
-
-    def test_confirmed_presentation_delete_forbidden_for_organizer(self):
-        """確定済みLTは主催者側の個別削除を拒否する"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time=time(21, 0),
-            confirmed_duration=60,
-            schedule_confirmed_at=timezone.now(),
-            progress=VketParticipation.Progress.REHEARSAL,
-        )
-        pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='確定済みLT',
-            theme='テーマ',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        response = self.client.post(
-            reverse(
-                'vket:presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-
-        self.assertEqual(response.status_code, 403)
-        self.assertTrue(VketPresentation.objects.filter(pk=pres.pk).exists())
-
-    def test_published_presentation_delete_forbidden_for_organizer(self):
-        """公開済みLTは主催者側の個別削除を拒否する"""
-        self.client.login(username='owner_user', password='testpass123')
-        self._set_active_community()
-        event = Event.objects.filter(community=self.community).first()
-        detail = EventDetail.objects.create(
-            event=event,
-            detail_type='LT',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            progress=VketParticipation.Progress.APPLIED,
-        )
-        pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='公開済みLT',
-            theme='テーマ',
-            published_event_detail=detail,
-        )
-
-        response = self.client.post(
-            reverse(
-                'vket:presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-
-        self.assertEqual(response.status_code, 403)
-        self.assertTrue(VketPresentation.objects.filter(pk=pres.pk).exists())
-        self.assertTrue(EventDetail.objects.filter(pk=detail.pk).exists())
-
-    def test_presentation_delete_forbidden_for_non_member(self):
-        """コミュニティに所属しないユーザーはLTを削除できない"""
-        User.objects.create_user(
-            user_name='non_member_user',
-            email='nonmember@example.com',
-            password='testpass123',
-        )
-        self.client.login(username='non_member_user', password='testpass123')
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            progress=VketParticipation.Progress.APPLIED,
-        )
-        pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='テスト',
-            status=VketPresentation.Status.DRAFT,
-        )
-        response = self.client.post(
-            reverse(
-                'vket:presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-        self.assertEqual(response.status_code, 403)
-        self.assertTrue(VketPresentation.objects.filter(pk=pres.pk).exists())
-
-
-class VketManageViewsTests(TestCase):
-    def setUp(self):
-        self.client = Client()
-        cache.clear()
-        self.superuser = User.objects.create_superuser(
-            user_name='admin_user',
-            email='admin@example.com',
-            password='adminpass123',
-        )
-        self.normal_user = User.objects.create_user(
-            user_name='normal_user',
-            email='normal@example.com',
-            password='testpass123',
-        )
-
-        self.community1 = Community.objects.create(
-            name='集会A',
-            status='approved',
-            frequency='毎週',
-        )
-        self.community2 = Community.objects.create(
-            name='集会B',
-            status='approved',
-            frequency='毎週',
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-manage-test',
-            name='Vket 2026 Summer 技術学術WEEK',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-        )
-
-        # 公開済みイベント（published_event）を使う
-        self.event1 = Event.objects.create(
-            community=self.community1,
-            date=today,
-            start_time='21:00',
-            duration=60,
-            weekday='Tue',
-        )
-        self.event2 = Event.objects.create(
-            community=self.community2,
-            date=today,
-            start_time='21:30',
-            duration=60,
-            weekday='Tue',
-        )
-
-        self.participation1 = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community1,
-            published_event=self.event1,
-            confirmed_date=today,
-            confirmed_start_time='21:00',
-            confirmed_duration=60,
-        )
-
-        self.participation2 = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community2,
-            published_event=self.event2,
-            confirmed_date=today,
-            confirmed_start_time='21:30',
-            confirmed_duration=60,
-        )
-
-    def tearDown(self):
-        cache.clear()
-
-    def test_manage_page_requires_staff(self):
-        """管理画面はstaff権限が必要"""
-        self.client.login(username='normal_user', password='testpass123')
-        response = self.client.get(reverse('vket:manage', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 403)
-
-    def test_manage_page_shows_collaboration(self):
-        """管理画面にコラボ名と集会名が表示される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        response = self.client.get(reverse('vket:manage', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.collaboration.name)
-        self.assertContains(response, self.community1.name)
-        self.assertContains(response, 'name="lifecycle"')
-        self.assertContains(response, '不参加')
-        self.assertContains(response, '辞退')
-
-    def test_manage_schedule_page_shows_overlap_warning(self):
-        """日程重複がある場合にoverlap_warningsがセットされる"""
-        self.client.login(username='admin_user', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_schedule', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.context['overlap_warnings'])
-
-    def test_manage_schedule_shows_confirmed_without_published_event(self):
-        """confirmed_dateがあればpublished_eventなしでも日程表に表示される"""
-        community3 = Community.objects.create(name='集会C', status='approved', frequency='毎週')
-        today = timezone.localdate()
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=community3,
-            confirmed_date=today,
-            confirmed_start_time='22:00',
-            confirmed_duration=60,
-        )
-
-        self.client.login(username='admin_user', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_schedule', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        rows = response.context['rows']
-        communities = [r['participation'].community.name for r in rows]
-        self.assertIn('集会C', communities)
-
-    def test_manage_schedule_shows_requested_without_confirmed_schedule(self):
-        """未確定でも希望日程があれば管理日程表に表示される"""
-        community3 = Community.objects.create(name='集会C', status='approved', frequency='毎週')
-        today = timezone.localdate()
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=community3,
-            requested_date=today,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-
-        self.client.login(username='admin_user', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_schedule', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        rows = response.context['rows']
-        requested_row = next(
-            r for r in rows if r['participation'].community.name == '集会C'
-        )
-        self.assertFalse(requested_row['is_confirmed'])
-        self.assertEqual(requested_row['date'], today)
-        self.assertEqual(requested_row['start_time'], time(22, 0))
-        self.assertContains(response, '集会C')
-        self.assertContains(response, '申請中')
-
-    def test_manage_schedule_page_marks_lt_slot(self):
-        """LT詳細があるスロットにlt_timesが設定される"""
-        EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-
-        self.client.login(username='admin_user', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_schedule', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        slots = response.context['slots']
-        expected_idx = next(i for i, s in enumerate(slots) if s.start == time(21, 30))
-
-        rows = response.context['rows']
-        row = next(r for r in rows if r['participation'].pk == self.participation1.pk)
-        lt_indices = [i for i, cell in enumerate(row['cells']) if cell.get('lt_times')]
-        self.assertEqual(lt_indices, [expected_idx])
-        self.assertEqual(row['cells'][expected_idx]['lt_times'], [time(21, 30)])
-
-    def test_manage_participation_update_sets_confirmed_fields(self):
-        """ManageParticipationUpdateViewが確定日程・progressを正しくセットする"""
-        self.client.login(username='admin_user', password='adminpass123')
-        # 未確定の参加を作成してテスト
-        new_participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会C', status='approved', frequency='毎週'),
-            requested_date=self.collaboration.period_start,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-        new_date = self.collaboration.period_start + timedelta(days=1)
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': new_participation.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '22:00',
-                'confirmed_duration': '60',
-                'admin_note': '確定しました',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        new_participation.refresh_from_db()
-        self.assertEqual(new_participation.confirmed_date, new_date)
-        self.assertEqual(new_participation.confirmed_start_time.strftime('%H:%M'), '22:00')
-        self.assertEqual(new_participation.confirmed_duration, 60)
-        self.assertEqual(new_participation.admin_note, '確定しました')
-        self.assertTrue(new_participation.schedule_adjusted_by_admin)
-        self.assertEqual(new_participation.progress, VketParticipation.Progress.REHEARSAL)
-        self.assertIsNotNone(new_participation.schedule_confirmed_at)
-        self.assertIsNotNone(new_participation.published_event_id)
-
-    def test_manage_participation_update_reuses_existing_event(self):
-        """確定日程に一致する既存Eventがあれば公開イベントとして再利用する"""
-        self.client.login(username='admin_user', password='adminpass123')
-        community = Community.objects.create(name='集会C', status='approved', frequency='毎週')
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-        existing_event = Event.objects.create(
-            community=community,
-            date=self.collaboration.period_start,
-            start_time='22:00',
-            duration=60,
-            weekday='Tue',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='既存Event登壇者',
-            theme='既存Eventテーマ',
-            requested_start_time='22:30',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': participation.pk,
-                },
-            ),
-            data={
-                'confirmed_date': self.collaboration.period_start.isoformat(),
-                'confirmed_start_time': '22:00',
-                'confirmed_duration': '60',
-            },
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        participation.refresh_from_db()
-        presentation.refresh_from_db()
-        self.assertEqual(participation.published_event_id, existing_event.pk)
-        self.assertEqual(Event.objects.filter(community=community).count(), 1)
-        self.assertEqual(presentation.published_event_detail.event_id, existing_event.pk)
-        self.assertEqual(presentation.published_event_detail.start_time.strftime('%H:%M'), '22:30')
-
-    def test_publication_sync_date_update_does_not_create_tombstone(self):
-        """Vket運営同期の日付更新はユーザー例外として記録しない"""
-        new_date = self.collaboration.period_start + timedelta(days=1)
-        self.participation1.confirmed_date = new_date
-        self.participation1.save(update_fields=['confirmed_date'])
-
-        sync_participation_publication(self.participation1)
-
-        self.event1.refresh_from_db()
-        self.assertEqual(self.event1.date, new_date)
-        self.assertEqual(
-            self.event1.weekday,
-            ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'][new_date.weekday()],
-        )
-        self.assertFalse(
-            EventOccurrenceTombstone.objects.filter(
-                community=self.community1,
-            ).exists()
-        )
-
-    def test_sync_participation_publication_sets_applicant_on_new_detail(self):
-        """公開同期で新規EventDetailへVket申請者を設定する。"""
-        self.participation1.applied_by = self.normal_user
-        self.participation1.save(update_fields=['applied_by'])
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='Vket発表者',
-            theme='Vketテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        sync_participation_publication(self.participation1)
-
-        presentation.refresh_from_db()
-        self.assertEqual(presentation.published_event_detail.applicant, self.normal_user)
-
-    def test_sync_participation_publication_backfills_applicant_on_existing_detail(self):
-        """再同期で既存EventDetailの未設定申請者を補完する。"""
-        self.participation1.applied_by = self.normal_user
-        self.participation1.save(update_fields=['applied_by'])
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='Vket発表者',
-            theme='Vketテーマ',
-            start_time='21:00',
-            duration=30,
-            status='approved',
-        )
-        VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='Vket発表者',
-            theme='Vketテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        sync_participation_publication(self.participation1)
-
-        detail.refresh_from_db()
-        self.assertEqual(detail.applicant, self.normal_user)
-
-    def test_sync_participation_publication_keeps_applicant_when_applied_by_is_none(self):
-        """Vket申請者が未設定でも既存EventDetailの申請者を保持する。"""
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='既存発表者',
-            theme='既存テーマ',
-            start_time='21:00',
-            duration=30,
-            status='approved',
-            applicant=self.normal_user,
-        )
-        VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='既存発表者',
-            theme='既存テーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        sync_participation_publication(self.participation1)
-
-        detail.refresh_from_db()
-        self.assertEqual(detail.applicant, self.normal_user)
-
-    def test_vket_applicant_backfill_uses_first_ordered_presentation(self):
-        """移行処理は表示順が先のVket発表に紐づく申請者を設定する。"""
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='移行対象',
-            theme='移行テーマ',
-            start_time='21:00',
-            duration=30,
-            status='approved',
-        )
-        self.participation1.applied_by = self.normal_user
-        self.participation1.save(update_fields=['applied_by'])
-        self.participation2.applied_by = self.superuser
-        self.participation2.save(update_fields=['applied_by'])
-        VketPresentation.objects.create(
-            participation=self.participation1,
-            published_event_detail=detail,
-            order=1,
-        )
-        VketPresentation.objects.create(
-            participation=self.participation2,
-            published_event_detail=detail,
-            order=0,
-        )
-
-        migration = import_module('vket.migrations.0010_backfill_vket_event_detail_applicant')
-        migration.backfill_vket_event_detail_applicant(apps, None)
-
-        detail.refresh_from_db()
-        self.assertEqual(detail.applicant, self.superuser)
-
-    def test_manage_participation_update_approves_existing_pending_detail(self):
-        """既存のpending EventDetailは確定時にapprovedへ同期される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='承認前登壇者',
-            theme='承認前テーマ',
-            start_time='21:30',
-            duration=30,
-            status='pending',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='承認後登壇者',
-            theme='承認後テーマ',
-            requested_start_time='21:45',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'confirmed_date': self.collaboration.period_start.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-            },
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        presentation.refresh_from_db()
-        detail.refresh_from_db()
-        self.assertEqual(detail.status, 'approved')
-        self.assertEqual(detail.speaker, '承認後登壇者')
-        self.assertEqual(detail.theme, '承認後テーマ')
-        self.assertEqual(detail.start_time.strftime('%H:%M'), '21:45')
-        self.assertEqual(presentation.published_event_detail_id, detail.pk)
-
-    def test_manage_participation_update_sets_lifecycle_without_schedule(self):
-        """管理画面から参加状態だけを不参加に変更できる"""
-        self.client.login(username='admin_user', password='adminpass123')
-        new_participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会C', status='approved', frequency='毎週'),
-            requested_date=self.collaboration.period_start,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': new_participation.pk,
-                },
-            ),
-            data={
-                'lifecycle': VketParticipation.Lifecycle.DECLINED,
-                'admin_note': '今回は不参加',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        new_participation.refresh_from_db()
-        self.assertEqual(new_participation.lifecycle, VketParticipation.Lifecycle.DECLINED)
-        self.assertEqual(new_participation.admin_note, '今回は不参加')
-        self.assertFalse(new_participation.schedule_adjusted_by_admin)
-        self.assertIsNone(new_participation.schedule_confirmed_at)
-
-    def test_manage_participation_update_declined_clears_published_event(self):
-        """公開済み参加を不参加にすると公開イベント連携を解除する"""
-        self.client.login(username='admin_user', password='adminpass123')
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'lifecycle': VketParticipation.Lifecycle.DECLINED,
-                'admin_note': '公開後に不参加',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        self.participation1.refresh_from_db()
-        presentation.refresh_from_db()
-        self.assertEqual(self.participation1.lifecycle, VketParticipation.Lifecycle.DECLINED)
-        self.assertIsNone(self.participation1.published_event_id)
-        self.assertIsNone(presentation.published_event_detail_id)
-        self.assertTrue(Event.objects.filter(pk=self.event1.pk).exists())
-        self.assertFalse(EventDetail.objects.filter(pk=detail.pk).exists())
-
-    def test_manage_participation_update_requires_schedule_for_active(self):
-        """参加中で保存する場合は確定日程が必須"""
-        self.client.login(username='admin_user', password='adminpass123')
-        new_participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会C', status='approved', frequency='毎週'),
-        )
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': new_participation.pk,
-                },
-            ),
-            data={
-                'lifecycle': VketParticipation.Lifecycle.ACTIVE,
-                'admin_note': '日程なし',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        new_participation.refresh_from_db()
-        self.assertIsNone(new_participation.confirmed_date)
-        self.assertFalse(new_participation.schedule_adjusted_by_admin)
-
-    def test_manage_participation_update_publishes_confirmed_presentation(self):
-        """確定ボタンは発表時刻を保存しEventDetailへ公開同期する"""
-        self.client.login(username='admin_user', password='adminpass123')
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会C', status='approved', frequency='毎週'),
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time='21:00',
-            confirmed_duration=60,
-        )
-        presentation = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='公開前登壇者',
-            theme='公開前テーマ',
-            requested_start_time='21:30',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-        new_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': participation.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-                'admin_note': '',
-                f'pres_{presentation.pk}_start_time': '22:15',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        presentation.refresh_from_db()
-        participation.refresh_from_db()
-        self.assertEqual(presentation.confirmed_start_time.strftime('%H:%M'), '22:15')
-        self.assertIsNotNone(participation.published_event_id)
-        self.assertIsNotNone(presentation.published_event_detail_id)
-        self.assertEqual(presentation.published_event_detail.event_id, participation.published_event_id)
-        self.assertEqual(presentation.published_event_detail.start_time.strftime('%H:%M'), '22:15')
-        self.assertEqual(presentation.published_event_detail.status, 'approved')
-
-    def test_manage_participation_update_updates_published_presentation_start_time(self):
-        """公開済み発表の開始時刻はVketPresentationとEventDetailに同期される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-        cache_key = get_index_view_cache_key()
-        cache.set(cache_key, {'upcoming_event_details': ['stale']}, 60)
-        new_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-                'admin_note': '',
-                f'pres_{presentation.pk}_start_time': '22:15',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        presentation.refresh_from_db()
-        detail.refresh_from_db()
-        self.assertEqual(presentation.confirmed_start_time.strftime('%H:%M'), '22:15')
-        self.assertEqual(detail.start_time.strftime('%H:%M'), '22:15')
-        self.assertIsNone(cache.get(cache_key))
-
-    def test_manage_participation_update_rejects_foreign_presentation(self):
-        """別参加の発表開始時刻は更新できない"""
-        self.client.login(username='admin_user', password='adminpass123')
-        foreign_detail = EventDetail.objects.create(
-            event=self.event2,
-            detail_type='LT',
-            speaker='別参加登壇者',
-            theme='別参加テーマ',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        foreign_presentation = VketPresentation.objects.create(
-            participation=self.participation2,
-            order=0,
-            speaker='別参加登壇者',
-            theme='別参加テーマ',
-            confirmed_start_time='21:30',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=foreign_detail,
-        )
-        new_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-                'admin_note': '',
-                f'pres_{foreign_presentation.pk}_start_time': '23:00',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        foreign_presentation.refresh_from_db()
-        foreign_detail.refresh_from_db()
-        self.assertEqual(foreign_presentation.confirmed_start_time.strftime('%H:%M'), '21:30')
-        self.assertEqual(foreign_detail.start_time.strftime('%H:%M'), '21:30')
-
-    def test_manage_participation_update_repairs_foreign_event_detail(self):
-        """同参加の発表に紐づく別イベントのEventDetailは公開イベントへ付け替える"""
-        self.client.login(username='admin_user', password='adminpass123')
-        foreign_detail = EventDetail.objects.create(
-            event=self.event2,
-            detail_type='LT',
-            speaker='別イベント登壇者',
-            theme='別イベントテーマ',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='不整合登壇者',
-            theme='不整合テーマ',
-            confirmed_start_time='21:30',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=foreign_detail,
-        )
-        cache_key = get_index_view_cache_key()
-        cache.set(cache_key, {'upcoming_event_details': ['stale']}, 60)
-        new_date = self.collaboration.period_start
-        response = self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-                'admin_note': '',
-                f'pres_{presentation.pk}_start_time': '23:00',
-            },
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        presentation.refresh_from_db()
-        foreign_detail.refresh_from_db()
-        self.assertEqual(presentation.confirmed_start_time.strftime('%H:%M'), '23:00')
-        self.assertEqual(foreign_detail.event_id, self.participation1.published_event_id)
-        self.assertEqual(foreign_detail.start_time.strftime('%H:%M'), '23:00')
-        self.assertEqual(foreign_detail.status, 'approved')
-        self.assertIsNone(cache.get(cache_key))
-
-    def test_manage_update_confirms_draft_presentations(self):
-        """確定ボタン押下でDRAFTのLTがCONFIRMEDに一括更新される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        new_community = Community.objects.create(
-            name='集会D', status='approved', frequency='毎週',
-        )
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=new_community,
-            requested_date=self.collaboration.period_start,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-        draft_pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='DRAFT登壇者',
-            theme='DRAFTテーマ',
-            status=VketPresentation.Status.DRAFT,
-        )
-        confirmed_pres = VketPresentation.objects.create(
-            participation=participation,
-            order=1,
-            speaker='確定済み登壇者',
-            theme='確定済みテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        new_date = self.collaboration.period_start + timedelta(days=1)
-        self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': participation.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '22:00',
-                'confirmed_duration': '60',
-            },
-        )
-
-        draft_pres.refresh_from_db()
-        confirmed_pres.refresh_from_db()
-        self.assertEqual(draft_pres.status, VketPresentation.Status.CONFIRMED)
-        self.assertEqual(confirmed_pres.status, VketPresentation.Status.CONFIRMED)
-
-    def test_manage_update_ignores_posted_start_time_for_initial_draft_presentation(self):
-        """DRAFT発表の開始時刻は同一POSTで送られても更新しない"""
-        self.client.login(username='admin_user', password='adminpass123')
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会E', status='approved', frequency='毎週'),
-            requested_date=self.collaboration.period_start,
-            requested_start_time='22:00',
-            requested_duration=60,
-        )
-        draft_pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='DRAFT登壇者',
-            theme='DRAFTテーマ',
-            status=VketPresentation.Status.DRAFT,
-        )
-
-        new_date = self.collaboration.period_start + timedelta(days=1)
-        self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': participation.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '22:00',
-                'confirmed_duration': '60',
-                f'pres_{draft_pres.pk}_start_time': '22:15',
-            },
-        )
-
-        draft_pres.refresh_from_db()
-        self.assertEqual(draft_pres.status, VketPresentation.Status.CONFIRMED)
-        self.assertIsNone(draft_pres.confirmed_start_time)
-
-    def test_manage_update_creates_event_detail_for_new_presentation(self):
-        """確定ボタンでpublished_event_detailがないCONFIRMED LTにEventDetailが作成される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        # published_event_detail がない DRAFT のLTを追加
-        pres = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=1,
-            speaker='新規登壇者',
-            theme='新規テーマ',
-            requested_start_time='21:45',
-            status=VketPresentation.Status.DRAFT,
-        )
-        cache_key = get_index_view_cache_key()
-        cache.set(cache_key, {'upcoming_event_details': ['stale']}, 60)
-        new_date = self.collaboration.period_start + timedelta(days=1)
-        self.client.post(
-            reverse(
-                'vket:manage_participation_update',
-                kwargs={
-                    'pk': self.collaboration.pk,
-                    'participation_id': self.participation1.pk,
-                },
-            ),
-            data={
-                'confirmed_date': new_date.isoformat(),
-                'confirmed_start_time': '21:00',
-                'confirmed_duration': '60',
-            },
-        )
-        pres.refresh_from_db()
-        self.assertEqual(pres.status, VketPresentation.Status.CONFIRMED)
-        self.assertIsNotNone(pres.published_event_detail)
-        self.assertEqual(pres.published_event_detail.speaker, '新規登壇者')
-        self.assertEqual(pres.published_event_detail.event_id, self.event1.id)
-        self.assertIsNone(cache.get(cache_key))
-
-    def test_manage_page_shows_draft_badge(self):
-        """管理画面でDRAFTのLTに「申請中」バッジが表示される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='DRAFT登壇者',
-            theme='DRAFTテーマ',
-            status=VketPresentation.Status.DRAFT,
-        )
-        response = self.client.get(
-            reverse('vket:manage', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'DRAFT登壇者')
-        self.assertContains(response, '申請中')
-        self.assertContains(response, 'badge bg-warning')
-
-    def test_manage_page_shows_time_input_for_rehearsal_confirmed_presentation(self):
-        """公開前のCONFIRMED発表には開始時刻入力欄を表示する"""
-        self.client.login(username='admin_user', password='adminpass123')
-        participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=Community.objects.create(name='集会C', status='approved', frequency='毎週'),
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time='21:00',
-            confirmed_duration=60,
-        )
-        draft_pres = VketPresentation.objects.create(
-            participation=participation,
-            order=0,
-            speaker='申請中登壇者',
-            theme='申請中テーマ',
-            status=VketPresentation.Status.DRAFT,
-        )
-        confirmed_pres = VketPresentation.objects.create(
-            participation=participation,
-            order=1,
-            speaker='確定済み登壇者',
-            theme='確定済みテーマ',
-            requested_start_time='21:30',
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        response = self.client.get(
-            reverse('vket:manage', kwargs={'pk': self.collaboration.pk})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, f'name="pres_{confirmed_pres.pk}_start_time"')
-        self.assertContains(response, 'value="21:30"')
-        self.assertNotContains(response, f'name="pres_{draft_pres.pk}_start_time"')
-
-    def test_manage_page_prefers_event_detail_time_for_published_presentation(self):
-        """公開済み発表の入力初期値は公開EventDetailの時刻を優先する"""
-        self.client.login(username='admin_user', password='adminpass123')
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            start_time='22:15',
-            duration=30,
-            status='approved',
-        )
-        presentation = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='公開済み登壇者',
-            theme='公開済みテーマ',
-            confirmed_start_time='21:30',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        response = self.client.get(
-            reverse('vket:manage', kwargs={'pk': self.collaboration.pk})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, f'name="pres_{presentation.pk}_start_time"')
-        self.assertContains(response, 'value="22:15"')
-
-    def test_manage_page_shows_lt_time_badge(self):
-        """管理画面でLT時間バッジが表示される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        # プレゼンテーションとEventDetailを作成
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='テスト登壇者',
-            theme='テストテーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-
-        response = self.client.get(
-            reverse('vket:manage', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, '21:30')
-        self.assertContains(response, 'badge bg-info')
-
-    def test_manage_presentation_delete_removes_presentation_and_event_detail(self):
-        """管理者がLTを削除するとVketPresentationとEventDetailが両方削除される"""
-        self.client.login(username='admin_user', password='adminpass123')
-        detail = EventDetail.objects.create(
-            event=self.event1,
-            detail_type='LT',
-            start_time='21:30',
-            duration=30,
-            status='approved',
-        )
-        pres = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='削除テスト登壇者',
-            theme='削除テーマ',
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-        response = self.client.post(
-            reverse(
-                'vket:manage_presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-        self.assertEqual(response.status_code, 302)
-        self.assertFalse(VketPresentation.objects.filter(pk=pres.pk).exists())
-        self.assertFalse(EventDetail.objects.filter(pk=detail.pk).exists())
-
-    def test_manage_presentation_delete_requires_staff(self):
-        """一般ユーザー（非staff）はLTを削除できない"""
-        self.client.login(username='normal_user', password='testpass123')
-        pres = VketPresentation.objects.create(
-            participation=self.participation1,
-            order=0,
-            speaker='テスト',
-            status=VketPresentation.Status.DRAFT,
-        )
-        response = self.client.post(
-            reverse(
-                'vket:manage_presentation_delete',
-                kwargs={'pk': self.collaboration.pk, 'presentation_id': pres.pk},
-            ),
-        )
-        self.assertEqual(response.status_code, 403)
-        self.assertTrue(VketPresentation.objects.filter(pk=pres.pk).exists())
-
-
-class VketStaffAccessTests(TestCase):
-    """is_staff=True, is_superuser=False のユーザーで管理機能にアクセスできることを検証."""
-
-    def setUp(self):
-        self.client = Client()
-        self.staff_user = User.objects.create_user(
-            user_name='staff_user',
-            email='staff@example.com',
-            password='staffpass123',
-            is_staff=True,
-        )
-
-        self.community = Community.objects.create(
-            name='Staff集会',
-            status='approved',
-            frequency='毎週',
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-staff-test',
-            name='Vket Staff Test',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-        )
-        self.draft_collaboration = VketCollaboration.objects.create(
-            slug='vket-staff-draft',
-            name='Vket Staff Draft',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.DRAFT,
-        )
-
-        self.event = Event.objects.create(
-            community=self.community,
-            date=today,
-            start_time='21:00',
-            duration=60,
-        )
-        self.participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            published_event=self.event,
-            confirmed_date=today,
-            confirmed_start_time='21:00',
-            confirmed_duration=60,
-        )
-
-        self.notice = VketNotice.objects.create(
-            collaboration=self.collaboration,
-            title='Staff通知テスト',
-            body='テスト本文',
-            created_by=self.staff_user,
-        )
-
-    def test_staff_can_access_manage_page(self):
-        """staffユーザーが管理画面にアクセスできる"""
-        self.client.login(username='staff_user', password='staffpass123')
-        response = self.client.get(reverse('vket:manage', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-
-    def test_staff_can_see_draft_collaboration(self):
-        """staffユーザーにDRAFTコラボが表示される"""
-        self.client.login(username='staff_user', password='staffpass123')
-        response = self.client.get(reverse('vket:list'))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.draft_collaboration.name)
-
-    def test_staff_can_access_manage_notice_list(self):
-        """staffユーザーが管理用お知らせ一覧にアクセスできる"""
-        self.client.login(username='staff_user', password='staffpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-
-    def test_staff_sees_admin_link_on_participation_status(self):
-        """staffユーザーの参加状況ページに管理リンクが表示される"""
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.staff_user,
-            role=CommunityMember.Role.OWNER,
-        )
-        self.client.login(username='staff_user', password='staffpass123')
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-
-        response = self.client.get(reverse('vket:status', kwargs={'pk': self.collaboration.pk}))
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, '管理')
-
-
-class BuildDiscordMentionsTests(TestCase):
-    """_build_discord_mentionsのユニットテスト."""
-
-    def setUp(self):
-        from allauth.socialaccount.models import SocialAccount
-
-        self.user1 = User.objects.create_user(
-            user_name='user1', email='u1@example.com', password='pass',
-        )
-        self.user2 = User.objects.create_user(
-            user_name='user2', email='u2@example.com', password='pass',
-        )
-        self.user3 = User.objects.create_user(
-            user_name='user3', email='u3@example.com', password='pass',
-        )
-        SocialAccount.objects.create(user=self.user1, provider='discord', uid='111')
-        SocialAccount.objects.create(user=self.user2, provider='discord', uid='222')
-        SocialAccount.objects.create(user=self.user3, provider='discord', uid='333')
-
-        community1 = Community.objects.create(name='集会1', status='approved', frequency='毎週')
-        community2 = Community.objects.create(name='集会2', status='approved', frequency='毎週')
-        community3 = Community.objects.create(name='集会3', status='approved', frequency='毎週')
-        today = timezone.localdate()
-        self.collab = VketCollaboration.objects.create(
-            slug='mention-test', name='テスト', period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=3),
-            lt_deadline=today + timedelta(days=5),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-        )
-
-        self.p1 = VketParticipation.objects.create(
-            collaboration=self.collab, community=community1,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-            applied_by=self.user1, progress=VketParticipation.Progress.APPLIED,
-        )
-        self.p2 = VketParticipation.objects.create(
-            collaboration=self.collab, community=community2,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-            applied_by=self.user2, progress=VketParticipation.Progress.STAGE_REGISTERED,
-        )
-        self.p3 = VketParticipation.objects.create(
-            collaboration=self.collab, community=community3,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-            applied_by=self.user3, progress=VketParticipation.Progress.REHEARSAL,
-        )
-
-    def _call(self):
-        from vket.views import ManageView
-        parts = VketParticipation.objects.filter(collaboration=self.collab)
-        return ManageView._build_discord_mentions(parts)
-
-    def test_stage_not_registered_contains_applied(self):
-        """APPLIED状態の参加者がstage_not_registeredメンションに含まれる."""
-        result = self._call()
-        self.assertIn('<@111>', result['stage_not_registered'])
-        self.assertNotIn('<@222>', result['stage_not_registered'])
-
-    def test_lt_not_registered_contains_stage_registered(self):
-        """STAGE_REGISTERED状態の参加者がlt_not_registeredメンションに含まれる."""
-        result = self._call()
-        self.assertIn('<@222>', result['lt_not_registered'])
-        self.assertNotIn('<@111>', result['lt_not_registered'])
-
-    def test_rehearsal_not_in_individual_groups(self):
-        """REHEARSAL状態の参加者は個別グループに含まれずallのみに含まれる."""
-        result = self._call()
-        self.assertIn('<@333>', result['all'])
-        self.assertNotIn('<@333>', result['stage_not_registered'])
-        self.assertNotIn('<@333>', result['lt_not_registered'])
-        self.assertNotIn('<@333>', result['not_applied'])
-
-    def test_all_contains_everyone(self):
-        """全参加者メンションに全員が含まれる."""
-        result = self._call()
-        self.assertIn('<@111>', result['all'])
-        self.assertIn('<@222>', result['all'])
-        self.assertIn('<@333>', result['all'])
-
-
-class VketNoticeTests(TestCase):
-    """お知らせ・ACK機能のテスト"""
-
-    def setUp(self):
-        self.client = Client()
-        self.superuser = User.objects.create_superuser(
-            user_name='admin_user2',
-            email='admin2@example.com',
-            password='adminpass123',
-        )
-        self.owner = User.objects.create_user(
-            user_name='owner_user2',
-            email='owner2@example.com',
-            password='testpass123',
-        )
-        self.community = Community.objects.create(
-            name='テスト集会',
-            status='approved',
-            frequency='毎週',
-        )
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.owner,
-            role=CommunityMember.Role.OWNER,
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-notice-test',
-            name='お知らせテスト',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.SCHEDULING,
-        )
-        self.participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-        )
-        self.notice = VketNotice.objects.create(
-            collaboration=self.collaboration,
-            title='テストお知らせ',
-            body='テスト本文',
-            requires_ack=True,
-            target_scope=VketNotice.TargetScope.ALL_PARTICIPANTS,
-            created_by=self.superuser,
-        )
-
-    def test_ack_notice_view_get_does_not_mark_acknowledged(self):
-        """AckNoticeView GET はプレビュー表示のみ（DB書き換えなし）"""
-        receipt = VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-        )
-        self.assertIsNone(receipt.acknowledged_at)
-
-        response = self.client.get(
-            reverse('vket:ack_notice', kwargs={'ack_token': str(receipt.ack_token)})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        receipt.refresh_from_db()
-        self.assertIsNone(receipt.acknowledged_at)  # GETでは変更されない
-        self.assertFalse(response.context['already_acked'])
-
-    def test_ack_notice_view_post_marks_acknowledged(self):
-        """AckNoticeView POST でreceiptのacknowledged_atがセットされる"""
-        receipt = VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-        )
-        self.assertIsNone(receipt.acknowledged_at)
-
-        response = self.client.post(
-            reverse('vket:ack_notice', kwargs={'ack_token': str(receipt.ack_token)})
-        )
-        self.assertEqual(response.status_code, 200)
-
-        receipt.refresh_from_db()
-        self.assertIsNotNone(receipt.acknowledged_at)
-        self.assertTrue(response.context['already_acked'])
-
-    def test_ack_notice_view_shows_already_acked(self):
-        """2回目のアクセスはalready_acked=Trueになる"""
-        receipt = VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-        response = self.client.get(
-            reverse('vket:ack_notice', kwargs={'ack_token': str(receipt.ack_token)})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.context['already_acked'])
-
-    def test_ack_notice_view_links_back_to_open_notice_detail(self):
-        """ACK済み画面から対象のお知らせ詳細を開いた一覧へ戻れる"""
-        receipt = VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        response = self.client.get(
-            reverse('vket:ack_notice', kwargs={'ack_token': str(receipt.ack_token)})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        detail_url = (
-            f"{reverse('vket:notice_list', kwargs={'pk': self.collaboration.pk})}"
-            f"?open={self.notice.pk}"
-        )
-        self.assertContains(response, detail_url)
-        self.assertContains(response, 'お知らせ一覧で詳細を見る')
-
-    def test_notice_list_view_requires_login(self):
-        """お知らせ一覧はログイン必須"""
-        response = self.client.get(
-            reverse('vket:notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        # ログイン画面にリダイレクト
-        self.assertEqual(response.status_code, 302)
-
-    def test_notice_list_view_shows_acked_notice_detail(self):
-        """参加者側一覧はACK済みでもお知らせ本文を開ける"""
-        self.notice.body = '1行目の詳細\n\n"引用" と & 記号を含む本文'
-        self.notice.save(update_fields=['body'])
-        VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        self.client.login(username='owner_user2', password='testpass123')
-        response = self.client.get(
-            reverse('vket:notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.notice.title)
-        self.assertContains(response, '確認済み')
-        self.assertContains(response, f'data-bs-target="#notice-{self.notice.pk}"')
-        self.assertContains(response, f'id="notice-{self.notice.pk}"')
-        self.assertContains(response, '1行目の詳細')
-        self.assertContains(response, '引用')
-
-    def test_manage_notice_list_requires_staff(self):
-        """管理用お知らせ一覧はstaff権限が必要"""
-        self.client.login(username='owner_user2', password='testpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 403)
-
-    def test_manage_notice_list_shows_notice(self):
-        """管理用お知らせ一覧にnoticeが表示される"""
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.notice.title)
-
-    def test_manage_notice_list_shows_acked_notice_detail_modal(self):
-        """管理一覧はACK済みでも全文詳細モーダルを表示する"""
-        self.notice.body = '管理向け詳細本文\n\n"引用" と & 記号を含む本文'
-        self.notice.save(update_fields=['body'])
-        VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, f'noticeDetailModal{self.notice.pk}')
-        self.assertContains(response, '管理向け詳細本文')
-        self.assertContains(response, '引用')
-        self.assertContains(response, '1/1確認')
-        self.assertContains(response, '作成日時')
-
-    def test_status_page_latest_notice_links_to_open_detail(self):
-        """参加状況ページの最新お知らせから対象詳細を開ける"""
-        VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        self.client.login(username='owner_user2', password='testpass123')
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-
-        self.assertEqual(response.status_code, 200)
-        detail_url = (
-            f"{reverse('vket:notice_list', kwargs={'pk': self.collaboration.pk})}"
-            f"?open={self.notice.pk}"
-        )
-        self.assertContains(response, detail_url, count=2)
-        self.assertContains(response, '詳細')
-
-    def test_notice_create_auto_generates_receipts(self):
-        """お知らせ作成時にactive参加者分のReceiptが自動生成される"""
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_notice_create', kwargs={'pk': self.collaboration.pk}),
-            data={
-                'title': '自動生成テスト',
-                'body': 'テスト本文',
-                'target_scope': 'all',
-                'requires_ack': 'on',
-            },
-        )
-        self.assertEqual(response.status_code, 302)
-        notice = VketNotice.objects.get(title='自動生成テスト')
-        receipts = VketNoticeReceipt.objects.filter(notice=notice)
-        self.assertEqual(receipts.count(), 1)
-        self.assertEqual(receipts.first().participation, self.participation)
-
-    def test_manage_notice_update_success(self):
-        """superuserがお知らせのタイトルと本文を編集できる"""
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_notice_update', kwargs={
-                'pk': self.collaboration.pk,
-                'notice_id': self.notice.pk,
-            }),
-            data={'title': '更新タイトル', 'body': '更新本文'},
-        )
-        self.assertEqual(response.status_code, 302)
-        self.notice.refresh_from_db()
-        self.assertEqual(self.notice.title, '更新タイトル')
-        self.assertEqual(self.notice.body, '更新本文')
-
-    def test_manage_notice_update_requires_staff(self):
-        """一般ユーザー（非staff）はお知らせを編集できない"""
-        self.client.login(username='owner_user2', password='testpass123')
-        response = self.client.post(
-            reverse('vket:manage_notice_update', kwargs={
-                'pk': self.collaboration.pk,
-                'notice_id': self.notice.pk,
-            }),
-            data={'title': '不正更新', 'body': '不正本文'},
-        )
-        self.assertEqual(response.status_code, 403)
-        self.notice.refresh_from_db()
-        self.assertEqual(self.notice.title, 'テストお知らせ')
-
-    def test_manage_notice_update_validates_required_fields(self):
-        """タイトル・本文が空の場合はエラー"""
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_notice_update', kwargs={
-                'pk': self.collaboration.pk,
-                'notice_id': self.notice.pk,
-            }),
-            data={'title': '', 'body': ''},
-        )
-        self.assertEqual(response.status_code, 302)
-        self.notice.refresh_from_db()
-        self.assertEqual(self.notice.title, 'テストお知らせ')
-
-    def test_manage_notice_update_rejects_other_collaboration_notice(self):
-        """他のcollaborationのお知らせは編集できない（404）"""
-        other_collab = VketCollaboration.objects.create(
-            slug='other-collab',
-            name='別コラボ',
-            period_start=timezone.localdate(),
-            period_end=timezone.localdate() + timedelta(days=7),
-            registration_deadline=timezone.localdate() + timedelta(days=1),
-            lt_deadline=timezone.localdate() + timedelta(days=3),
-            phase=VketCollaboration.Phase.SCHEDULING,
-        )
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_notice_update', kwargs={
-                'pk': other_collab.pk,
-                'notice_id': self.notice.pk,
-            }),
-            data={'title': '不正更新', 'body': '不正本文'},
-        )
-        self.assertEqual(response.status_code, 404)
-
-    def test_manage_notice_list_unacked_mentions_role(self):
-        """未ACKのコミュニティのロールメンションがコンテキストに含まれる"""
-        self.community.discord_mention_type = Community.DiscordMentionType.ROLE
-        self.community.discord_mention_role_id = '111222333'
-        self.community.save()
-
-        VketNoticeReceipt.objects.create(
-            notice=self.notice, participation=self.participation
-        )
-
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        stat = response.context['notice_stats'][0]
-        self.assertIn('<@&111222333>', stat['unacked_mentions'])
-
-    def test_manage_notice_list_unacked_mentions_users(self):
-        """未ACKのコミュニティのユーザーメンションがコンテキストに含まれる"""
-        self.community.discord_mention_type = Community.DiscordMentionType.USERS
-        self.community.discord_mention_user_ids = ['aaa', 'bbb']
-        self.community.save()
-
-        VketNoticeReceipt.objects.create(
-            notice=self.notice, participation=self.participation
-        )
-
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        stat = response.context['notice_stats'][0]
-        self.assertIn('<@aaa>', stat['unacked_mentions'])
-        self.assertIn('<@bbb>', stat['unacked_mentions'])
-
-    def test_manage_notice_list_acked_not_in_mentions(self):
-        """ACK済みのコミュニティはメンションに含まれない"""
-        self.community.discord_mention_type = Community.DiscordMentionType.ROLE
-        self.community.discord_mention_role_id = '999888777'
-        self.community.save()
-
-        VketNoticeReceipt.objects.create(
-            notice=self.notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        stat = response.context['notice_stats'][0]
-        self.assertEqual(stat['unacked_mentions'], [])
-
-    def test_manage_notice_list_unacked_mentions_fallback_discord_id(self):
-        """メンション未設定のコミュニティはメンバーのDiscord IDでメンション生成"""
-        from allauth.socialaccount.models import SocialAccount
-        SocialAccount.objects.create(
-            user=self.owner, provider='discord', uid='123456789'
-        )
-        VketNoticeReceipt.objects.create(
-            notice=self.notice, participation=self.participation
-        )
-
-        self.client.login(username='admin_user2', password='adminpass123')
-        response = self.client.get(
-            reverse('vket:manage_notice_list', kwargs={'pk': self.collaboration.pk})
-        )
-        stat = response.context['notice_stats'][0]
-        self.assertIn('<@123456789>', stat['unacked_mentions'])
-
-
-class VketParticipationStatusTests(TestCase):
-    """ParticipationStatusView のテスト（未確認お知らせバナー等）"""
-
-    def setUp(self):
-        self.client = Client()
-        self.owner = User.objects.create_user(
-            user_name='status_owner',
-            email='status_owner@example.com',
-            password='testpass123',
-        )
-        self.community = Community.objects.create(
-            name='ステータステスト集会',
-            status='approved',
-            frequency='毎週',
-        )
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.owner,
-            role=CommunityMember.Role.OWNER,
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-status-test',
-            name='ステータステスト',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.SCHEDULING,
-        )
-        self.participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-        )
-
-    def _set_active_community(self):
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-
-    def test_status_page_shows_unacked_count(self):
-        """未確認のrequires_ackお知らせがあるとバナーに件数が表示される"""
-        superuser = User.objects.create_superuser(
-            user_name='su_status', email='su_status@example.com', password='p',
-        )
-        notice = VketNotice.objects.create(
-            collaboration=self.collaboration,
-            title='要確認',
-            body='本文',
-            requires_ack=True,
-            target_scope=VketNotice.TargetScope.ALL_PARTICIPANTS,
-            created_by=superuser,
-        )
-        VketNoticeReceipt.objects.create(
-            notice=notice, participation=self.participation,
-        )
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.context['unacked_count'], 1)
-        self.assertContains(response, '未確認のお知らせ')
-
-    def test_status_page_unacked_count_zero_when_all_acked(self):
-        """全レシートがACK済みならバナーは表示されない"""
-        superuser = User.objects.create_superuser(
-            user_name='su_status2', email='su_status2@example.com', password='p',
-        )
-        notice = VketNotice.objects.create(
-            collaboration=self.collaboration,
-            title='確認済み',
-            body='本文',
-            requires_ack=True,
-            target_scope=VketNotice.TargetScope.ALL_PARTICIPANTS,
-            created_by=superuser,
-        )
-        VketNoticeReceipt.objects.create(
-            notice=notice,
-            participation=self.participation,
-            acknowledged_at=timezone.now(),
-        )
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.context['unacked_count'], 0)
-        self.assertNotContains(response, '未確認のお知らせ')
-
-    def test_status_page_without_participation(self):
-        """participation がない場合 unacked_count=0"""
-        other_user = User.objects.create_user(
-            user_name='no_part_user', email='nopart@example.com', password='testpass123',
-        )
-        other_community = Community.objects.create(
-            name='別の集会', status='approved', frequency='毎週',
-        )
-        CommunityMember.objects.create(
-            community=other_community,
-            user=other_user,
-            role=CommunityMember.Role.OWNER,
-        )
-
-        self.client.login(username='no_part_user', password='testpass123')
-        session = self.client.session
-        session['active_community_id'] = other_community.id
-        session.save()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.context['unacked_count'], 0)
-        self.assertNotContains(response, '未確認のお知らせ')
-
-    def test_stage_register_advances_progress(self):
-        """ステージ登録POSTで progress が STAGE_REGISTERED に進む"""
-        self.participation.progress = VketParticipation.Progress.APPLIED
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.post(
-            reverse('vket:stage_register', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        self.participation.refresh_from_db()
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.STAGE_REGISTERED)
-        self.assertIsNotNone(self.participation.stage_registered_at)
-
-    def test_stage_register_requires_applied_participation(self):
-        """未申請ではステージ登録できない"""
-        self.participation.progress = VketParticipation.Progress.NOT_APPLIED
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.post(
-            reverse('vket:stage_register', kwargs={'pk': self.collaboration.pk}),
-            follow=True,
-        )
-        self.assertEqual(response.status_code, 200)
-
-        self.participation.refresh_from_db()
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.NOT_APPLIED)
-        self.assertIsNone(self.participation.stage_registered_at)
-        self.assertContains(response, 'ステージ登録は参加申込み後に行ってください。')
-
-    def test_stage_register_records_rehearsal_without_rewinding_progress(self):
-        """REHEARSAL では進捗を戻さずステージ登録日時だけ記録する"""
-        self.participation.progress = VketParticipation.Progress.REHEARSAL
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.post(
-            reverse('vket:stage_register', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        self.participation.refresh_from_db()
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.REHEARSAL)
-        self.assertIsNotNone(self.participation.stage_registered_at)
-
-    def test_stage_register_rejects_after_period_end(self):
-        """開催期間終了後はステージ登録日時を記録しない"""
-        today = timezone.localdate()
-        self.collaboration.period_start = today - timedelta(days=7)
-        self.collaboration.period_end = today - timedelta(days=1)
-        self.collaboration.save(update_fields=['period_start', 'period_end'])
-        self.participation.progress = VketParticipation.Progress.APPLIED
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.post(
-            reverse('vket:stage_register', kwargs={'pk': self.collaboration.pk}),
-            follow=True,
-        )
-        self.assertEqual(response.status_code, 200)
-
-        self.participation.refresh_from_db()
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.APPLIED)
-        self.assertIsNone(self.participation.stage_registered_at)
-        self.assertContains(response, '開催期間終了後はステージ登録を記録できません。')
-
-    def test_stage_register_rejects_duplicate_registration(self):
-        """二重登録では既存の登録日時を上書きしない"""
-        registered_at = timezone.now() - timedelta(days=1)
-        self.participation.progress = VketParticipation.Progress.REHEARSAL
-        self.participation.stage_registered_at = registered_at
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.post(
-            reverse('vket:stage_register', kwargs={'pk': self.collaboration.pk}),
-            follow=True,
-        )
-        self.assertEqual(response.status_code, 200)
-
-        self.participation.refresh_from_db()
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.REHEARSAL)
-        self.assertEqual(self.participation.stage_registered_at, registered_at)
-        self.assertContains(response, 'ステージ登録は既に完了しています。')
-
-    def test_status_page_shows_stage_banner(self):
-        """申請済み時にステージ登録バナーが表示される"""
-        self.participation.progress = VketParticipation.Progress.APPLIED
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'Vketステージ登録')
-
-    def test_status_page_hides_stage_register_section_after_registration(self):
-        """登録済み後はステージ登録セクションを表示しない"""
-        self.participation.progress = VketParticipation.Progress.STAGE_REGISTERED
-        self.participation.stage_registered_at = timezone.now()
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertFalse(response.context['stage_register_open'])
-        self.assertNotContains(response, 'Vketステージ登録')
-        self.assertNotContains(response, '登録済み（')
-
-    def test_status_page_shows_stage_register_button_for_rehearsal_within_period(self):
-        """期間内のREHEARSAL未登録では登録完了ボタンを表示する"""
-        self.participation.progress = VketParticipation.Progress.REHEARSAL
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertTrue(response.context['stage_register_open'])
-        self.assertContains(response, '登録完了')
-
-    def test_status_page_hides_stage_register_button_after_period_end(self):
-        """期間終了後の未登録は登録ボタンを出さず未登録バッジだけ表示する"""
-        today = timezone.localdate()
-        self.collaboration.period_start = today - timedelta(days=7)
-        self.collaboration.period_end = today - timedelta(days=1)
-        self.collaboration.save(update_fields=['period_start', 'period_end'])
-        self.participation.progress = VketParticipation.Progress.REHEARSAL
-        self.participation.save()
-
-        self.client.login(username='status_owner', password='testpass123')
-        self._set_active_community()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertFalse(response.context['stage_register_open'])
-        self.assertNotContains(response, '登録完了')
-        self.assertContains(response, '未登録')
-
-
-class VketStatusRedirectViewTests(TestCase):
-    """VketStatusRedirectView のテスト"""
-
-    def setUp(self):
-        self.client = Client()
-        self.user = User.objects.create_user(
-            user_name='redirect_user',
-            email='redirect@example.com',
-            password='testpass123',
-        )
-        self.superuser = User.objects.create_superuser(
-            user_name='redirect_admin',
-            email='redirect_admin@example.com',
-            password='adminpass123',
-        )
-        self.community = Community.objects.create(
-            name='リダイレクトテスト集会',
-            status='approved',
-            frequency='毎週',
-        )
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.user,
-            role=CommunityMember.Role.OWNER,
-        )
-
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-redirect-test',
-            name='リダイレクトテスト',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.ENTRY_OPEN,
-        )
-
-    def test_redirect_requires_login(self):
-        """未ログインで302（ログインページへ）"""
-        response = self.client.get(reverse('vket:status_redirect'))
-        self.assertEqual(response.status_code, 302)
-        self.assertIn('/login/', response.url)
-
-    def test_redirect_to_latest_collaboration(self):
-        """ログインユーザーで最新コラボのstatusへリダイレクト"""
-        self.client.login(username='redirect_user', password='testpass123')
-        response = self.client.get(reverse('vket:status_redirect'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.url,
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk}),
-        )
-
-    def test_redirect_to_list_when_only_archived(self):
-        """ARCHIVEDのみの場合、公開一覧にリダイレクト"""
-        self.collaboration.phase = VketCollaboration.Phase.ARCHIVED
-        self.collaboration.save()
-
-        self.client.login(username='redirect_user', password='testpass123')
-        response = self.client.get(reverse('vket:status_redirect'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse('vket:list'))
-
-    def test_redirect_skips_draft_for_non_superuser(self):
-        """非superuserでDRAFTのみの場合、公開一覧にリダイレクト"""
-        self.collaboration.phase = VketCollaboration.Phase.DRAFT
-        self.collaboration.save()
-
-        self.client.login(username='redirect_user', password='testpass123')
-        response = self.client.get(reverse('vket:status_redirect'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(response.url, reverse('vket:list'))
-
-    def test_superuser_can_access_draft(self):
-        """superuserはDRAFTコラボにもリダイレクトされる"""
-        self.collaboration.phase = VketCollaboration.Phase.DRAFT
-        self.collaboration.save()
-
-        self.client.login(username='redirect_admin', password='adminpass123')
-        response = self.client.get(reverse('vket:status_redirect'))
-        self.assertEqual(response.status_code, 302)
-        self.assertEqual(
-            response.url,
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk}),
-        )
-
-    def test_status_page_shows_collaboration_dropdown(self):
-        """コラボ切替ドロップダウンが status ページに表示される"""
-        today = timezone.localdate()
-        second = VketCollaboration.objects.create(
-            slug='vket-2026-redirect-test-2',
-            name='リダイレクトテスト2',
-            period_start=today - timedelta(days=30),
-            period_end=today - timedelta(days=23),
-            registration_deadline=today - timedelta(days=29),
-            lt_deadline=today - timedelta(days=27),
-            phase=VketCollaboration.Phase.LOCKED,
-        )
-
-        self.client.login(username='redirect_user', password='testpass123')
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, '切替')
-        self.assertContains(response, second.name)
-
-    def test_status_page_shows_admin_link_for_superuser(self):
-        """superuserで管理者リンク（管理画面）が表示される"""
-        # superuserにもCommunityMembershipが必要
-        CommunityMember.objects.create(
-            community=self.community,
-            user=self.superuser,
-            role=CommunityMember.Role.OWNER,
-        )
-        self.client.login(username='redirect_admin', password='adminpass123')
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-
-        # participationを作成（管理者リンクはアクションボタンエリアに表示）
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-        )
-
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, 'fa-gear')
-
-    def test_status_page_hides_admin_link_for_normal_user(self):
-        """一般ユーザーには管理者リンクが表示されない"""
-        self.client.login(username='redirect_user', password='testpass123')
-        session = self.client.session
-        session['active_community_id'] = self.community.id
-        session.save()
-
-        VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-        )
-
-        response = self.client.get(
-            reverse('vket:status', kwargs={'pk': self.collaboration.pk})
-        )
-        self.assertEqual(response.status_code, 200)
-        self.assertNotContains(response, 'fa-gear')
-
-
-class VketPublishViewTests(TestCase):
-    """ManagePublishViewのテスト"""
-
-    def setUp(self):
-        self.client = Client()
-        cache.clear()
-        self.superuser = User.objects.create_superuser(
-            user_name='admin_pub',
-            email='admin_pub@example.com',
-            password='adminpass123',
-        )
-        self.community = Community.objects.create(
-            name='公開テスト集会',
-            status='approved',
-            frequency='毎週',
-        )
-        today = timezone.localdate()
-        self.collaboration = VketCollaboration.objects.create(
-            slug='vket-2026-publish-test',
-            name='公開テスト',
-            period_start=today,
-            period_end=today + timedelta(days=7),
-            registration_deadline=today + timedelta(days=1),
-            lt_deadline=today + timedelta(days=3),
-            phase=VketCollaboration.Phase.LOCKED,
-        )
-        self.participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=self.community,
-            lifecycle=VketParticipation.Lifecycle.ACTIVE,
-            confirmed_date=today,
-            confirmed_start_time='21:00',
-            confirmed_duration=60,
-        )
-
-    def tearDown(self):
-        cache.clear()
-
-    def test_publish_creates_event_and_updates_participation(self):
-        """公開処理でEventが作成されpublished_eventが紐づく"""
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        self.participation.refresh_from_db()
-        self.assertIsNotNone(self.participation.published_event_id)
-        self.assertEqual(self.participation.progress, VketParticipation.Progress.DONE)
-
-        event = self.participation.published_event
-        self.assertEqual(event.community, self.community)
-        self.assertEqual(event.start_time.strftime('%H:%M'), '21:00')
-        self.assertEqual(event.duration, 60)
-        self.assertEqual(
-            event.weekday,
-            ['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'][event.date.weekday()],
-        )
-
-    def test_publish_skips_declined_participation(self):
-        """不参加の参加情報は公開同期の対象外になる"""
-        declined_community = Community.objects.create(
-            name='不参加テスト集会',
-            status='approved',
-            frequency='毎週',
-        )
-        declined_participation = VketParticipation.objects.create(
-            collaboration=self.collaboration,
-            community=declined_community,
-            lifecycle=VketParticipation.Lifecycle.DECLINED,
-            confirmed_date=self.collaboration.period_start,
-            confirmed_start_time='22:00',
-            confirmed_duration=60,
-        )
-        VketPresentation.objects.create(
-            participation=declined_participation,
-            order=0,
-            speaker='不参加登壇者',
-            theme='不参加テーマ',
-            requested_start_time=time(22, 30),
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        declined_participation.refresh_from_db()
-        self.assertIsNone(declined_participation.published_event_id)
-        self.assertFalse(Event.objects.filter(community=declined_community).exists())
-        self.assertFalse(EventDetail.objects.filter(speaker='不参加登壇者').exists())
-
-    def test_publish_is_forbidden_when_not_locked(self):
-        """LOCKEDフェーズ以外では公開処理が403になる"""
-        self.collaboration.phase = VketCollaboration.Phase.SCHEDULING
-        self.collaboration.save()
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-        )
-        self.assertEqual(response.status_code, 403)
-
-    def test_publish_is_idempotent(self):
-        """公開処理を2回実行しても同じEventが使われ、重複作成されない"""
-        from event.models import Event
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        url = reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk})
-
-        # 1回目
-        self.client.post(url)
-        self.participation.refresh_from_db()
-        first_event_id = self.participation.published_event_id
-        self.assertIsNotNone(first_event_id)
-
-        # 2回目
-        self.client.post(url)
-        self.participation.refresh_from_db()
-        # 同じEventが使われていること
-        self.assertEqual(self.participation.published_event_id, first_event_id)
-        # Eventが重複作成されていないこと
-        self.assertEqual(Event.objects.filter(community=self.community).count(), 1)
-
-    def test_lt_start_time_flows_to_event_detail(self):
-        """requested_start_time が EventDetail.start_time に反映される"""
-        # プレゼンテーションに requested_start_time を設定（CONFIRMED で公開対象にする）
-        VketPresentation.objects.create(
-            participation=self.participation,
-            order=0,
-            speaker='テスト登壇者',
-            theme='テストテーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.CONFIRMED,
-        )
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-        self.assertEqual(response.status_code, 302)
-
-        self.participation.refresh_from_db()
-        event = self.participation.published_event
-        self.assertIsNotNone(event)
-
-        # EventDetail が作成され、start_time に requested_start_time が使われていること
-        detail = EventDetail.objects.get(event=event)
-        self.assertEqual(detail.start_time.strftime('%H:%M'), '21:30')
-
-    def test_publish_skips_draft_presentation(self):
-        """申請中の発表は公開処理でEventDetailにしない"""
-        VketPresentation.objects.create(
-            participation=self.participation,
-            order=0,
-            speaker='申請中登壇者',
-            theme='申請中テーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.DRAFT,
-        )
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        self.participation.refresh_from_db()
-        self.assertIsNotNone(self.participation.published_event_id)
-        self.assertFalse(EventDetail.objects.filter(speaker='申請中登壇者').exists())
-
-    def test_publish_clears_index_cache_when_updating_existing_event_detail(self):
-        """公開処理で既存EventDetailをbulk updateした場合もトップページキャッシュを削除する"""
-        event = Event.objects.create(
-            community=self.community,
-            date=self.collaboration.period_start,
-            start_time='21:00',
-            duration=60,
-            weekday='Fri',
-        )
-        self.participation.published_event = event
-        self.participation.save(update_fields=['published_event', 'updated_at'])
-        detail = EventDetail.objects.create(
-            event=event,
-            detail_type='LT',
-            speaker='更新前登壇者',
-            theme='更新前テーマ',
-            start_time='21:00',
-            status='approved',
-        )
-        VketPresentation.objects.create(
-            participation=self.participation,
-            order=0,
-            speaker='更新後登壇者',
-            theme='更新後テーマ',
-            requested_start_time=time(21, 30),
-            status=VketPresentation.Status.CONFIRMED,
-            published_event_detail=detail,
-        )
-        cache_key = get_index_view_cache_key()
-        cache.set(cache_key, {'upcoming_event_details': ['stale']}, 60)
-
-        self.client.login(username='admin_pub', password='adminpass123')
-        response = self.client.post(
-            reverse('vket:manage_publish', kwargs={'pk': self.collaboration.pk}),
-            follow=False,
-        )
-
-        self.assertEqual(response.status_code, 302)
-        detail.refresh_from_db()
-        self.assertEqual(detail.speaker, '更新後登壇者')
-        self.assertEqual(detail.start_time.strftime('%H:%M'), '21:30')
-        self.assertIsNone(cache.get(cache_key))
